+++ Tests/CustomDumpTests/DumpTests.swift
@@ -715,573 +715,6 @@ final class DumpTests: XCTestCase {
     )
   }
 
-  func testFoundation() {
-    var dump = ""
-
-    #if compiler(>=5.5) && !targetEnvironment(macCatalyst) && (os(iOS) || os(tvOS) || os(watchOS))
-      if #available(iOS 15, macOS 12, tvOS 15, watchOS 8, *) {
-        dump = ""
-        customDump(
-          try? AttributedString(markdown: "Hello, **Blob**!"),
-          to: &dump
-        )
-        XCTAssertNoDifference(
-          dump,
-          """
-          "Hello, Blob!"
-          """
-        )
-      }
-    #endif
-
-    // NB: `CFNumber` is unavailable on Linux
-    #if os(iOS) || os(macOS) || os(tvOS) || os(watchOS)
-      dump = ""
-      customDump(
-        42 as CFNumber,
-        to: &dump
-      )
-      XCTAssertNoDifference(
-        dump,
-        """
-        42
-        """
-      )
-    #endif
-
-    dump = ""
-    customDump(
-      Date(timeIntervalSince1970: 0),
-      to: &dump
-    )
-    XCTAssertNoDifference(
-      dump,
-      """
-      Date(1970-01-01T00:00:00.000Z)
-      """
-    )
-
-    #if compiler(>=5.4)
-      dump = ""
-      customDump(
-        NestedDate(date: Date(timeIntervalSince1970: 0)),
-        to: &dump
-      )
-      XCTAssertNoDifference(
-        dump,
-        """
-        NestedDate(date: Date(1970-01-01T00:00:00.000Z))
-        """
-      )
-    #endif
-
-    dump = ""
-    customDump(
-      Decimal(string: "1.23"),
-      to: &dump
-    )
-    XCTAssertNoDifference(
-      dump,
-      """
-      1.23
-      """
-    )
-
-    dump = ""
-    customDump(
-      [1, 2, 3] as NSArray,
-      to: &dump
-    )
-    XCTAssertNoDifference(
-      dump,
-      """
-      [
-        [0]: 1,
-        [1]: 2,
-        [2]: 3
-      ]
-      """
-    )
-
-    let attributedString = NSMutableAttributedString(string: "")
-    attributedString.append(NSAttributedString(string: "Hello, "))
-    attributedString.append(
-      NSAttributedString(string: "Blob", attributes: [.init(rawValue: "name"): true])
-    )
-    attributedString.append(NSAttributedString(string: "!"))
-    dump = ""
-    customDump(
-      attributedString,
-      to: &dump
-    )
-    XCTAssertNoDifference(
-      dump,
-      """
-      "Hello, Blob!"
-      """
-    )
-
-    let calendar = NSCalendar(calendarIdentifier: .gregorian)!
-    calendar.timeZone = TimeZone(secondsFromGMT: 0)!
-    dump = ""
-    customDump(
-      calendar,
-      to: &dump
-    )
-    XCTAssertNoDifference(
-      dump,
-      """
-      Calendar(
-        identifier: Calendar.Identifier.gregorian,
-        locale: Locale(),
-        timeZone: TimeZone(
-          identifier: "GMT",
-          abbreviation: "GMT",
-          secondsFromGMT: 0,
-          isDaylightSavingTime: false
-        ),
-        firstWeekday: 1,
-        minimumDaysInFirstWeek: 1
-      )
-      """
-    )
-
-    dump = ""
-    customDump(
-      NSCountedSet(array: [1, 2, 2, 3, 3, 3]),
-      to: &dump
-    )
-    XCTAssertNoDifference(
-      dump,
-      """
-      Set([
-        1,
-        2,
-        3
-      ])
-      """
-    )
-
-    dump = ""
-    customDump(
-      NSData(data: .init(repeating: 0, count: 4)),
-      to: &dump
-    )
-    XCTAssertNoDifference(
-      dump,
-      """
-      Data(4 bytes)
-      """
-    )
-
-    dump = ""
-    customDump(
-      NSDate(timeIntervalSince1970: 0),
-      to: &dump
-    )
-    XCTAssertNoDifference(
-      dump,
-      """
-      Date(1970-01-01T00:00:00.000Z)
-      """
-    )
-
-    dump = ""
-    customDump(
-      [1: "1", 2: "2", 3: "3"] as NSDictionary,
-      to: &dump
-    )
-    XCTAssertNoDifference(
-      dump,
-      """
-      [
-        1: "1",
-        2: "2",
-        3: "3"
-      ]
-      """
-    )
-
-    dump = ""
-    customDump(
-      NSError(
-        domain: "co.pointfree",
-        code: 42,
-        userInfo: [
-          NSLocalizedDescriptionKey: "An error occurred" as NSString
-        ]
-      ),
-      to: &dump
-    )
-    XCTAssertNoDifference(
-      dump,
-      """
-      NSError(
-        domain: "co.pointfree",
-        code: 42,
-        userInfo: [
-          "NSLocalizedDescription": "An error occurred"
-        ]
-      )
-      """
-    )
-
-    #if !os(Windows)
-      class SubclassedError: NSError {}
-
-      dump = ""
-      customDump(
-        SubclassedError(
-          domain: "co.pointfree",
-          code: 43,
-          userInfo: [
-            NSLocalizedDescriptionKey: "An error occurred" as NSString
-          ]
-        ),
-        to: &dump
-      )
-      XCTAssertNoDifference(
-        dump,
-        """
-        NSError(
-          domain: "co.pointfree",
-          code: 43,
-          userInfo: [
-            "NSLocalizedDescription": "An error occurred"
-          ]
-        )
-        """
-      )
-    #endif
-
-    enum BridgedError: Error {
-      case thisIsFine(Int)
-    }
-
-    dump = ""
-    customDump(BridgedError.thisIsFine(94) as NSError, to: &dump)
-    #if os(iOS) || os(macOS) || os(tvOS) || os(watchOS)
-      XCTAssertNoDifference(
-        dump,
-        """
-        DumpTests.BridgedError.thisIsFine(94)
-        """
-      )
-    #elseif compiler(>=5.4)
-      // Can't unwrap bridged Errors on Linux: https://bugs.swift.org/browse/SR-15191
-      XCTAssertNoDifference(
-        dump,
-        """
-        NSError(
-          domain: "CustomDumpTests.DumpTests.BridgedError",
-          code: 0,
-          userInfo: [:]
-        )
-        """
-      )
-    #endif
-
-    // NB: `NSException` is unavailable on Linux
-    #if os(iOS) || os(macOS) || os(tvOS) || os(watchOS)
-      dump = ""
-      customDump(
-        NSException(name: .genericException, reason: "Oops!", userInfo: nil),
-        to: &dump
-      )
-      XCTAssertNoDifference(
-        dump,
-        """
-        NSException(
-          name: NSGenericException,
-          reason: "Oops!",
-          userInfo: nil
-        )
-        """
-      )
-    #endif
-
-    // NB: `NSExpression` is unavailable on Linux
-    #if os(iOS) || os(macOS) || os(tvOS) || os(watchOS)
-      dump = ""
-      customDump(
-        NSExpression(format: "1 + 1"),
-        to: &dump
-      )
-      XCTAssertNoDifference(
-        dump,
-        """
-        1 + 1
-        """
-      )
-    #endif
-
-    dump = ""
-    customDump(
-      NSIndexPath(),
-      to: &dump
-    )
-    XCTAssertNoDifference(
-      dump,
-      """
-      []
-      """
-    )
-
-    dump = ""
-    customDump(
-      NSIndexSet(indexSet: [1, 2, 3, 5, 7]),
-      to: &dump
-    )
-    XCTAssertNoDifference(
-      dump,
-      """
-      IndexSet(
-        ranges: [
-          [0]: 1..<4,
-          [1]: 5..<6,
-          [2]: 7..<8
-        ]
-      )
-      """
-    )
-
-    dump = ""
-    customDump(
-      NSLocale(localeIdentifier: "en_US"),
-      to: &dump
-    )
-    XCTAssertNoDifference(
-      dump,
-      """
-      Locale(en_US)
-      """
-    )
-
-    dump = ""
-    customDump(
-      NSMeasurement(doubleValue: 42, unit: Unit(symbol: "kg")),
-      to: &dump
-    )
-    XCTAssertNoDifference(
-      dump,
-      """
-      Measurement(
-        value: 42.0,
-        unit: "kg"
-      )
-      """
-    )
-
-    dump = ""
-    customDump(
-      NSNotification(name: .init(rawValue: "co.pointfree"), object: nil, userInfo: nil),
-      to: &dump
-    )
-    XCTAssertNoDifference(
-      dump,
-      """
-      Notification(name: "co.pointfree")
-      """
-    )
-
-    dump = ""
-    customDump(
-      NSNull(),
-      to: &dump
-    )
-    XCTAssertNoDifference(
-      dump,
-      """
-      NSNull()
-      """
-    )
-
-    dump = ""
-    customDump(
-      NSNumber(booleanLiteral: true),
-      to: &dump
-    )
-    XCTAssertNoDifference(
-      dump,
-      """
-      1
-      """
-    )
-
-    dump = ""
-    customDump(
-      [1, 2, 3] as NSOrderedSet,
-      to: &dump
-    )
-    XCTAssertNoDifference(
-      dump,
-      """
-      [
-        [0]: 1,
-        [1]: 2,
-        [2]: 3
-      ]
-      """
-    )
-
-    dump = ""
-    customDump(
-      NSRange(0..<1),
-      to: &dump
-    )
-    XCTAssertNoDifference(
-      dump,
-      """
-      0..<1
-      """
-    )
-
-    dump = ""
-    customDump(
-      NSSet(array: [1, 2, 3]),
-      to: &dump
-    )
-    XCTAssertNoDifference(
-      dump,
-      """
-      Set([
-        1,
-        2,
-        3
-      ])
-      """
-    )
-
-    dump = ""
-    customDump(
-      NSTimeZone(forSecondsFromGMT: 0),
-      to: &dump
-    )
-    XCTAssertNoDifference(
-      dump,
-      """
-      TimeZone(
-        identifier: "GMT",
-        abbreviation: "GMT",
-        secondsFromGMT: 0,
-        isDaylightSavingTime: false
-      )
-      """
-    )
-
-    dump = ""
-    customDump(
-      NSURL(fileURLWithPath: "/tmp"),
-      to: &dump
-    )
-    #if os(Windows)
-      XCTAssertNoDifference(
-        dump,
-        """
-        URL(file:///tmp)
-        """
-      )
-    #else
-      XCTAssertNoDifference(
-        dump,
-        """
-        URL(file:///tmp/)
-        """
-      )
-    #endif
-
-    dump = ""
-    customDump(
-      NSURLComponents(string: "https://www.pointfree.co/login?redirect=episodes"),
-      to: &dump
-    )
-    XCTAssertNoDifference(
-      dump,
-      """
-      URLComponents(
-        scheme: "https",
-        host: "www.pointfree.co",
-        path: "/login",
-        queryItems: [
-          [0]: URLQueryItem(
-            name: "redirect",
-            value: "episodes"
-          )
-        ]
-      )
-      """
-    )
-
-    dump = ""
-    customDump(
-      NSURLQueryItem(name: "search", value: "composable architecture"),
-      to: &dump
-    )
-    XCTAssertNoDifference(
-      dump,
-      """
-      URLQueryItem(
-        name: "search",
-        value: "composable architecture"
-      )
-      """
-    )
-
-    dump = ""
-    let request = NSMutableURLRequest(url: URL(string: "https://www.pointfree.co")!)
-    request.addValue("text/html", forHTTPHeaderField: "Accept")
-    request.httpShouldUsePipelining = false
-    customDump(
-      request,
-      to: &dump
-    )
-    XCTAssertNoDifference(
-      dump,
-      """
-      URLRequest(
-        url: URL(https://www.pointfree.co),
-        cachePolicy: 0,
-        timeoutInterval: 60.0,
-        mainDocumentURL: nil,
-        networkServiceType: URLRequest.NetworkServiceType.default,
-        allowsCellularAccess: true,
-        httpMethod: "GET",
-        allHTTPHeaderFields: [
-          "Accept": "text/html"
-        ],
-        httpBody: nil,
-        httpBodyStream: nil,
-        httpShouldHandleCookies: true,
-        httpShouldUsePipelining: false
-      )
-      """
-    )
-
-    dump = ""
-    customDump(
-      NSUUID(uuidString: "deadbeef-dead-beef-dead-beefdeadbeef"),
-      to: &dump
-    )
-    XCTAssertNoDifference(
-      dump,
-      """
-      UUID(DEADBEEF-DEAD-BEEF-DEAD-BEEFDEADBEEF)
-      """
-    )
-
-    dump = ""
-    customDump(
-      URL(string: "https://www.pointfree.co/"),
-      to: &dump
-    )
-    XCTAssertNoDifference(
-      dump,
-      """
-      URL(https://www.pointfree.co/)
-      """
-    )
-  }
-
   #if canImport(CoreGraphics)
     func testCoreGraphics() {
       var dump = ""
